+++ upgrade/init-4.3.0.php
@@ -1,0 +1,47 @@
+<?php
+/**
+ * Copyright since 2007 PrestaShop SA and Contributors
+ * PrestaShop is an International Registered Trademark & Property of PrestaShop SA
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is bundled with this package in the file LICENSE.md.
+ * It is also available through the world-wide-web at this URL:
+ * https://opensource.org/licenses/OSL-3.0
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to license@prestashop.com so we can send you a copy immediately.
+ *
+ * @author    PrestaShop SA and Contributors <contact@prestashop.com>
+ * @copyright Since 2007 PrestaShop SA and Contributors
+ * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
+ */
+
+use Comfino\Main;
+
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+/**
+ * @return bool
+ */
+function upgrade_module_4_3_0(Comfino $module)
+{
+    if (!$module->checkEnvironment()) {
+        return false;
+    }
+
+    // Remove legacy config keys retired by the V3 paywall migration.
+    Configuration::deleteByName('COMFINO_PAYWALL_URL');
+    Configuration::deleteByName('COMFINO_SHOW_LOGO');
+
+    // Remove legacy paywall-init JS files (replaced by the inline V3 SDK bootstrap in payment.tpl).
+    @unlink(_PS_MODULE_DIR_ . $module->name . '/views/js/front/paywall-init.js');
+    @unlink(_PS_MODULE_DIR_ . $module->name . '/views/js/front/paywall-init.min.js');
+
+    Main::updateUpgradeLog('Upgrade script for 4.3.0 executed.');
+
+    return true;
+}